DM-12660 Handle long values for primary keys in herd entities and model-api objects

+++ herd-code/herd-app/src/test/java/org/finra/herd/app/security/SecurityUserWrapperTest.java
@@ -71,7 +71,7 @@ public class SecurityUserWrapperTest extends AbstractAppTest
                 businessObjectDefinitionDao.getBusinessObjectDefinitionByKey(new BusinessObjectDefinitionKey(NAMESPACE, BDEF_NAME));
 
             // Validate the newly created entity.
-            assertEquals(Integer.valueOf(businessObjectDefinition.getId()), businessObjectDefinitionEntity.getId());
+            assertEquals(Long.valueOf(businessObjectDefinition.getId()), businessObjectDefinitionEntity.getId());
             String expectedUsername = TrustedApplicationUserBuilder.TRUSTED_USER_ID;
             assertEquals(expectedUsername, businessObjectDefinitionEntity.getCreatedBy());
             assertEquals(expectedUsername, businessObjectDefinitionEntity.getUpdatedBy());
